moved batch and code classes to the domain.
