+++ src/web-app/domain/Batch.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace CodeJar.WebApp
+namespace CodeJar.Domain
 {
     public class Batch
     {
