--- src/web-app/domain/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace domain
-{
-    public class Class1
-    {
-    }
-}
