--- src/web-app/infrastructure/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace infrastructure
-{
-    public class Class1
-    {
-    }
-}
