+++ src/web-app/server/TableData.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Data;
+using CodeJar.Domain;
 
 namespace CodeJar.WebApp
 {
